feat: endpoint para listar todos os monitores

+++ src/Hippocampus.Domain/Diplomat/HttpIn/RecipientMonitorPostDto.cs
@@ -8,8 +8,8 @@ public class RecipientMonitorPostDto
 {
     public string Name { get; init; } = string.Empty;
     public MacAddress MacAddress { get; init; } = new();
-    public float MinHeight { get; init; }
-    public float MaxHeight { get; init; }
+    public int MinHeight { get; init; }
+    public int MaxHeight { get; init; }
     public RecipientType RecipientType { get; init; }
     public MacAddress? RecipientMonitorLinkedToMacAddress { get; init; } = new();
 }
